moved course views to a sub folder course update imports
