+++ src/views/course/CourseCreationPage.vue
